fx: move prefix command to slash 'admin tower', remove CHANGELOG.md
