+++ slash_commands/subcommands/admin/shop/delete.js
@@ -1,7 +1,7 @@
-const { createError, createLogs } = require("../../../util/envoiMsg");
+const { createError, createLogs } = require("../../../../util/envoiMsg");
 const { EmbedBuilder } = require("discord.js");
-const { YELLOW, NIGHT } = require("../../../data/colors.json");
-const { CHECK_MARK } = require("../../../data/emojis.json");
+const { YELLOW, NIGHT } = require("../../../../data/colors.json");
+const { CHECK_MARK } = require("../../../../data/emojis.json");
 const mongoose = require("mongoose");
 
 async function deleteItem(interaction, options) {
