+++ slash_commands/subcommands/admin/shop/index.js
