+++ slash_commands/subcommands/admin/tower/index.js
@@ -1,0 +1,7 @@
+const { start } = require("./start");
+const { stop } = require("./stop");
+
+module.exports = {
+    start,
+    stop,
+};
